FINAL/Latency_test/MODELE_YOLO_CATTLE/requirements_latency_test.txt
antlr4-python3-runtime==4.9.3
certifi==2025.1.31
charset-normalizer==3.4.1
contourpy==1.1.1
cycler==0.12.1
effdet==0.4.1
filelock==3.16.1
fonttools==4.57.0
fsspec==2025.3.0
gitdb==4.0.12
GitPython==3.1.44
huggingface-hub==0.30.2
idna==3.10
importlib_resources==6.4.5
Jinja2==3.1.6
kiwisolver==1.4.7
MarkupSafe==2.1.5
matplotlib==3.7.5
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
opencv-python==4.11.0.86
packaging==24.2
pandas==2.0.3
pillow==10.4.0
psutil==7.0.0
py-cpuinfo==9.0.0
pycocotools==2.0.7
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
safetensors==0.5.3
scipy==1.10.1
seaborn==0.13.2
six==1.17.0
smmap==5.0.2
sympy==1.13.3
thop==0.1.1.post2209072238
timm==1.0.15
torch==2.4.1
torchvision==0.19.1
tqdm==4.67.1
triton==3.0.0
typing_extensions==4.13.2
tzdata==2025.2
ultralytics==8.3.107
ultralytics-thop==2.0.14
urllib3==2.2.3
zipp==3.20.2
